LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animation-base-response-003-expected.txt


FAIL Identical elements are all responsive to font-size animation assert_equals: expected "15px" but got "1px"
